requirements.txt
streamlit
torch
torchvision
Pillow
numpy
grad-cam
opencv-python-headless